Fix #249: gate NameIndex rebuild on Name component count, not entity id
`NameIndex.generation` was tracking `world.next_entity_id()`, so
every entity spawn — named or not — bumped the generation and
forced a full HashMap rebuild on the next animation_system tick.
A 3000-entity cell load with only 500 named entities triggered a
pointless O(500) rebuild for the 2500 unnamed spawns.

Swap the generation proxy to the count of `Name` components:
`QueryRead::len()` is O(1) (reads the storage's element count), so
the check itself is free. Only spawns/despawns that actually touch
`Name` now invalidate the cache.

Changes:

- `byroredux/src/components.rs` — `NameIndex.generation` changed
  from `u32` to `usize` to match the count type. Initial sentinel
  becomes `usize::MAX`. Doc comment updated to explain the new
  semantic and call out the edge case where in-place `Name`
  replacement (no count change) would NOT invalidate the index —
  no engine code currently renames entities post-spawn.

- `byroredux/src/systems.rs` — the rebuild gate now compares
  `NameIndex.generation` against `world.query::<Name>().map(|q|
  q.len()).unwrap_or(0)` instead of `world.next_entity_id()`. The
  query is acquired twice per frame (once for the count check,
  once for the rebuild loop if needed); both acquisitions are
  O(1) read locks against the same storage and coexist cheaply.

Functional behavior:

- Unnamed-entity spawn: count unchanged → no rebuild (was
  unnecessary rebuild).
- Named-entity spawn: count incremented → rebuild (unchanged).
- Named-entity despawn: count decremented → rebuild (unchanged).
- First call (empty world): `usize::MAX != 0` → rebuild (same as
  before with `u32::MAX != 0`).

No new tests: the NameIndex is a private component indirectly
exercised by the animation system integration tests, and the
behavior only differs on the "unnamed spawn" path which is not
visible from outside.

Tests: cargo test — full workspace, 0 failures.

Audit: docs/audits/AUDIT_PERFORMANCE_2026-04-11.md PERF-04-11-L4.

+++ byroredux/src/components.rs
@@ -65,10 +65,26 @@ pub(crate) struct CellLightingRes {
 impl Resource for CellLightingRes {}
 
 /// Cached name→entity mapping for the animation system.
-/// Rebuilt only when the entity count changes (no per-frame allocations).
+///
+/// Rebuilt only when the count of `Name` components changes. Previously
+/// the generation tracked `world.next_entity_id()`, which forced a full
+/// rebuild on every entity spawn regardless of whether the spawn
+/// involved a `Name` — a 3000-entity cell load with only 500 named
+/// entities still triggered one rebuild on the next frame. Using the
+/// `Name` storage size as the generation means only spawns/despawns
+/// that actually touch `Name` invalidate the cache. See #249.
+///
+/// Edge case: in-place `Name` replacement (re-inserting `Name` on an
+/// existing entity without removing it first) does not change the
+/// count and therefore does not invalidate the index. No code in the
+/// engine currently renames entities after spawn, so this is not a
+/// concern today — add an explicit `invalidate()` call if that
+/// changes.
 pub(crate) struct NameIndex {
     pub(crate) map: HashMap<FixedString, EntityId>,
-    pub(crate) generation: u32,
+    /// Count of `Name` components seen at the last rebuild. `usize::MAX`
+    /// on a fresh index so the first comparison always rebuilds.
+    pub(crate) generation: usize,
 }
 impl Resource for NameIndex {}
 
@@ -76,7 +92,7 @@ impl NameIndex {
     pub(crate) fn new() -> Self {
         Self {
             map: HashMap::new(),
-            generation: u32::MAX, // Force rebuild on first use.
+            generation: usize::MAX, // Force rebuild on first use.
         }
     }
 }
